更新商家信息，管理员修改数据接口

+++ src/main/java/com/project/mall/controller/req/merchant/MerchantQueryProductByNameReq.java
@@ -16,5 +16,5 @@ public class MerchantQueryProductByNameReq {
      * 卖家ID
      */
 
-    private int MerchantID;
+    private Long MerchantID;
 }
